убрать лишние строки в коде класса BaseHttpHandler; убрать явное приведение типов в методе removeSubTaskId класса Epic;

+++ src/com/koval/kanban/model/Epic.java
@@ -25,7 +25,7 @@ public class Epic extends Task {
     }
 
     public void removeSubTaskId(int subTaskId) {
-        subTaskIdList.remove((Integer) subTaskId);
+        subTaskIdList.remove(subTaskId);
     }
 
     @Override
